+++ packages/full/index.js
@@ -1,9 +1,9 @@
 export * from "@croquet/worldcore-kernel";
-export * from "@croquet/worldcore-behavior";
 export * from "@croquet/worldcore-three";
 export * from "@croquet/worldcore-widget2";
 export * from "@croquet/worldcore-widget3";
 
+// export * from "@croquet/worldcore-behavior";
 // export * from "@croquet/worldcore-audio";
 // export * from "@croquet/worldcore-card";
 // export * from "@croquet/worldcore-rapier";